Simplifica validação de senha da sala

+++ src/resources/api.js
@@ -8,30 +8,33 @@ api.updateName = async (user, setUser, name) => {
     let hasError
     let missingName = false
 
-        try {
-            const res = await api.put(`/person/${user.person.id}`, {name}, {
-                    headers: {
-                        Authorization: 'Bearer ' + user.token    
-                    }
-            })
-            setUser({...user, person:{...user.person, name: res.data.name}})
-            localStorage.setItem('user', JSON.stringify(user))
-            hasError = false
+    if (user.person.name === name)
+        return hasError = false
+
+    try {
+        const res = await api.put(`/person/${user.person.id}`, {name}, {
+                headers: {
+                    Authorization: 'Bearer ' + user.token    
+                }
+        })
+        setUser({...user, person:{...user.person, name: res.data.name}})
+        localStorage.setItem('user', JSON.stringify(user))
+        hasError = false
+    }
+    catch (err) {
+
+        if (err.response.data.message === "name can't be empty")
+            missingName = true
+
+        if (err.response.data.message === "Failed to authenticate token") {
+            localStorage.clear('user')
+            setUser(null)
         }
-        catch (err) {
 
-            if (err.response.data.message === "name can't be empty")
-                missingName = true
+        hasError = true
+    }
 
-            if (err.response.data.message === "Failed to authenticate token") {
-                localStorage.clear('user')
-                setUser(null)
-            }
-
-            hasError = true
-        }
-
-        return hasError? {missingName}:false
+    return hasError? {missingName}:false
 }
